hdmi: dispdev: Kernel 3.0 migration changes
Add the kernel 3.0 migration changes for
dispdev and hdmi drivers

ST-Ericsson ID: 352334
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: NA

Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/29975

+++ drivers/misc/dispdev/dispdev.c
@@ -430,7 +430,7 @@ static int dispdev_dequeue_buffer(struct dispdev *dd)
 	return i;
 }
 
-int dispdev_ioctl(struct file *file, unsigned int cmd,
+long dispdev_ioctl(struct file *file, unsigned int cmd,
 							unsigned long arg)
 {
 	int ret;
